enhance blocker positioning logic

+++ app/src/main/java/com/example/racingcar/models/Blocker.kt
@@ -1,3 +1,3 @@
 package com.example.racingcar.models
 
-data class Blocker(val id: Int, val lanePosition: Int)
+data class Blocker(val id: Long, val lanePosition: Int)
